Added JavaPOS service interface generator.
The generator has been checked to generate all versions in the past
correctly (in case we need to regenerate them due to generally changed
header comments or so). The the newly generated code is not literally
equal (ordering and line endings changed and compiler optimization
applied), but it has been checked file by file that it is semantically
equivalent to the original. So it would be safe to regenerate the later
on.
However, the generator will normally be only used to generate service
interfaces of new JavaPOS versions (1.15 and above).