Added JavaPOS device control generator.
The generator has been checked to generate version 1.14 correctly. The
the newly generated code is not literally
equal (ordering and line endings changed and compiler optimization
applied), but it has been checked file by file that it is semantically
equivalent to the original. So it would be safe to use it for version
1.15.